audio policy: fix volume problem - DO NOT MERGE
merge from master:

"audio policy: fix volume change when disabling SCO

commit 76e97d39 introduced a regression by having setDeviceConnectionState()
force a device change on duplicated outputs even if the selected device is 0.
This also forces a device 0 for the two outputs it is duplicated to,
which may override a valid device selection on those outputs and apply
default volumes.

Bug 7429869."

Change-Id: I37dd66343c54eb27e420089edbda04444fa1e2c6